src/sync_images/python/requirements.txt
dotenv
pcloud
pyyaml